requirements.txt
einops
einops_exts
av
omegaconf
sentencepiece
protobuf
transformers
accelerate
decord
pandas 
scikit-learn
tqdm
wandb
huggingface-hub[cli]
pillow
open-clip-torch
# pip install flash-attn==2.7.3 --no-build-isolation
torchvision
peft
triton==2.3.1
deepspeed